EACH/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
supabase
langgraph
langgraph-swarm
langchain-openai
langchain-core
google-generativeai
openai
